refactor(frontend): extract the test run stat card
SummaryCard was a private function inside TestRunHeader, so the metric
tuning tab could not show its agreement the way a test run shows its pass
rate. Moved to components/common unchanged — same props, same markup —
and imported back where it was.
